fix: resolve layout and navigation nesting structure
- Move navigation and i18n providers to locale layout
- Separate layout styles into dedicated SCSS file
- Remove duplicate HTML structure
