+++ frontend/src/app/[locale]/layout.scss
@@ -1,0 +1,14 @@
+.layout-container-main {
+  box-sizing: border-box;
+  width: 100%;
+  min-height: 100vh;
+  overflow-y: auto;
+
+  @media (min-width: 601px) {
+    padding: 0 0 0 80px;
+  }
+
+  @media (max-width: 600px) {
+    padding: 64px 0 80px;
+  }
+}
