+++ frontend/src/app/[locale]/layout.tsx
@@ -1,47 +1,12 @@
-import { setRequestLocale } from 'next-intl/server';
-import { NextIntlClientProvider } from 'next-intl';
-import { getLocale, getMessages } from 'next-intl/server';
-import { notFound } from 'next/navigation';
-import { routing } from '@/i18n/routing';
-import type { Metadata } from 'next';
-import '../globals.scss';
+import Navigation from '@/components/Navigation/Navigation';
+import I18nProvider from './providers/i18n-provider';
+import './layout.scss';
 
-export const metadata: Metadata = {
-  title: 'HADA',
-  description: 'HADA'
-};
-
-const theme = 'light';
-
-export function generateStaticParams() {
-  return routing.locales.map((locale) => ({ locale }));
-}
-
-export default async function LocaleLayout({
-  children
-}: {
-  children: React.ReactNode;
-}) {
-  // 현재 locale 가져오기
-  const locale = await getLocale();
-  // Providing all messages to the client
-  const messages = await getMessages();
-
-  // Ensure that the incoming `locale` is valid
-  if (!routing.locales.includes(locale as any)) {
-    notFound();
-  }
-
-  // Enable static rendering
-  setRequestLocale(locale);
-
+export default async function LocaleLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang={locale}>
-      <body className={`antialiased ${theme}`}>
-        <NextIntlClientProvider messages={messages} >
-          {children}
-        </NextIntlClientProvider>
-      </body>
-    </html>
+    <I18nProvider>
+      <Navigation />
+      <div className="layout-container-main">{children}</div>
+    </I18nProvider>
   );
 }
