+++ frontend/src/app/[locale]/providers/i18n-provider.tsx
@@ -1,0 +1,7 @@
+import { NextIntlClientProvider } from 'next-intl';
+import { getMessages } from 'next-intl/server';
+
+export default async function I18nProvider({ children }: { children: React.ReactNode }) {
+  const messages = await getMessages();
+  return <NextIntlClientProvider messages={messages}>{children}</NextIntlClientProvider>;
+}
