+++ frontend/src/app/globals.scss
@@ -40,18 +40,3 @@ body {
   color: var(--foreground);
   font-family: var(--md-sys-typescale-body-large-font-family-name);
 }
-
-.layout-container-main {
-  box-sizing: border-box;
-  width: 100%;
-  min-height: 100vh;
-  overflow-y: auto;
-
-  @media (min-width: 601px) {
-    padding: 0 0 0 80px;
-  }
-
-  @media (max-width: 600px) {
-    padding: 64px 0 80px;
-  }
-}
