+++ frontend/src/app/layout.tsx
@@ -1,7 +1,10 @@
+import { setRequestLocale } from 'next-intl/server';
+import { getLocale } from 'next-intl/server';
+import { notFound } from 'next/navigation';
+import { routing } from '@/i18n/routing';
 import type { Metadata } from 'next';
-import './globals.scss';
 
-import Navigation from '@/components/Navigation/Navigation';
+import './globals.scss';
 
 export const metadata: Metadata = {
   title: 'HADA',
@@ -10,23 +13,31 @@ export const metadata: Metadata = {
 
 const theme = 'light';
 
-export default function RootLayout({
-  children
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+export function generateStaticParams() {
+  return routing.locales.map((locale) => ({ locale }));
+}
+
+export default async function LocaleLayout({ children }: { children: React.ReactNode }) {
+  // 현재 locale 가져오기
+  const locale = await getLocale();
+
+  // Ensure that the incoming `locale` is valid
+  if (!routing.locales.includes(locale as any)) {
+    notFound();
+  }
+
+  // Enable static rendering
+  setRequestLocale(locale);
+
   return (
-    <html lang="en">
+    <html lang={locale}>
       <head>
         <link
           href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
           rel="stylesheet"
         />
       </head>
-      <body className={`antialiased ${theme}`}>
-        <Navigation />
-        <main className="layout-container-main">{children}</main>
-      </body>
+      <body className={`antialiased ${theme}`}>{children}</body>
     </html>
   );
 }
